update test for fastani rule
- move all tests to single `.py`
- update function comparing fastani files
- update conftest.py
