+++ tests/conftest.py
@@ -103,18 +103,6 @@ def anim_nucmer_targets_delta(anim_nucmer_targets_delta_outdir):
     return [anim_nucmer_targets_delta_outdir / _.name for _ in reference_paths]
 
 
-# @pytest.fixture
-# def anim_nucmer_expected_targets_filter():
-#     """Target files for ANIm tests.
-
-#     These are paths to the output files we generated using
-#     nucmer for ANIm. We aim to use these to test if the filter
-#     files generated by snakemake workflow are the same.
-#     """
-
-#     return FIXTUREPATH / "anim" / "targets" / "filter"
-
-
 @pytest.fixture
 def fastani_targets_outdir():
     """Output directory for fastani snakemake tests.
@@ -138,7 +126,7 @@ def fastani_targets(fastani_targets_outdir):
 
 
 @pytest.fixture
-def fastani_expected_targets():
+def fastani_targets_indir():
     """Target files for fastANI tests.
 
     These are paths to the output files we generated using
